feat(UIKIT-1306,ui,Notify): Экспортирует тип для использования в классах
